pyproject.toml
[project]
name = "latvian-einvoice"
description = "Unified Python SDK for Latvia's e-adrese (VRAA VUS) and VID EDS e-Invoice API"
authors = [{ name = "guntisratkevics", email = "guntis.rat@gmail.com" }]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
  "zeep>=4.2",
  "requests>=2.31",
  "python-dateutil>=2.9",
  "cryptography>=42.0",
]
dynamic = ["version"]

[project.optional-dependencies]
async = ["httpx>=0.27"]

[project.urls]
Homepage = "https://github.com/guntisratkevics/eadrese"
Bug_Tracker = "https://github.com/guntisratkevics/eadrese/issues"

[tool.setuptools_scm]
version_scheme = "post-release"
local_scheme = "no-local-version"

[build-system]
requires = ["setuptools>=68", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"
